pyproject.toml
[tool.poetry]
name = "Beer Tap Dispenser Api by osw4l"
version = "0.1.0"
description = ""
authors = ["osw4l <ioswxd@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
drf-yasg = "^1.21.4"
factory-boy = "^3.2.1"

[tool.poetry.dev-dependencies]
coverage = "^6.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

